+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0 (12-7-2016)
+- auto_get = True
+- disable automatic discover cache
+- if you use auto_get = False must make a .get() to get the data in the class
+- Add support Python 2
+
 # 0.1.0 (11-7-2016)
 - Add collect property